+++ main.py
@@ -1,4 +1,84 @@
-'''Testing script'''
+'''Dataset preprocessing scripts'''
+
+
+def process_mim_gold_ner():
+    from pathlib import Path
+    import pandas as pd
+    from tqdm.auto import tqdm
+    import json
+    import re
+    from collections import defaultdict
+
+    conversion_dict = {
+        'O': 'O',
+        'B-Person': 'B-PER',
+        'I-Person': 'I-PER',
+        'B-Location': 'B-LOC',
+        'I-Location': 'I-LOC',
+        'B-Organization': 'B-ORG',
+        'I-Organization': 'I-ORG',
+        'B-Miscellaneous': 'B-MISC',
+        'I-Miscellaneous': 'I-MISC',
+        'B-Date': 'B-MISC',
+        'I-Date': 'I-MISC',
+        'B-Time': 'B-MISC',
+        'I-Time': 'I-MISC',
+        'B-Money': 'B-MISC',
+        'I-Money': 'I-MISC',
+        'B-Percent': 'B-MISC',
+        'I-Percent': 'I-MISC'
+    }
+
+    def get_df(path: Path):
+        lines = path.read_text().split('\n')
+        data_dict = defaultdict(list)
+        tokens = list()
+        tags = list()
+        for line in tqdm(lines):
+            if line != '':
+                token, tag = line.split('\t')
+                tag = conversion_dict[tag]
+                tokens.append(token)
+                tags.append(tag)
+            else:
+                doc = ' '.join(tokens)
+                doc = re.sub(' ([.,])', '\1', doc)
+                data_dict['doc'].append(doc)
+                data_dict['tokens'].append(tokens)
+                data_dict['ner_tags'].append(tags)
+                tokens = list()
+                tags = list()
+
+        return pd.DataFrame(data_dict)
+
+    def export_as_jsonl(df: pd.DataFrame, output_path: Path):
+        for idx, row in tqdm(list(df.iterrows())):
+            data_dict = dict(doc=row.doc,
+                             tokens=row.tokens,
+                             ner_tags=row.ner_tags)
+            json_line = json.dumps(data_dict)
+            with output_path.open('a') as f:
+                f.write(json_line)
+                if idx < len(df) - 1:
+                    f.write('\n')
+
+    input_dir = Path('datasets') / 'train-valid-test-split'
+    train_input_path = input_dir / 'train' / 'train'
+    val_input_path = input_dir / 'valid' / 'valid'
+    test_input_path = input_dir / 'test' / 'test'
+
+    output_dir = Path('datasets') / 'mim_gold_ner'
+    train_output_path = output_dir / 'train.jsonl'
+    test_output_path = output_dir / 'test.jsonl'
+
+    if not output_dir.exists():
+        output_dir.mkdir()
+
+    train_df = pd.concat((get_df(train_input_path), get_df(val_input_path)))
+    test_df = get_df(test_input_path)
+
+    export_as_jsonl(train_df, train_output_path)
+    export_as_jsonl(test_df, test_output_path)
 
 
 def process_fdt():
@@ -1585,4 +1665,4 @@ def process_absabank_imm():
 
 
 if __name__ == '__main__':
-    process_sdt()
+    process_mim_gold_ner()