+++ scandeval/benchmarks/__init__.py
@@ -23,6 +23,7 @@ from .idt_pos import IdtPosBenchmark  # noqa
 from .idt_dep import IdtDepBenchmark  # noqa
 from .fdt_pos import FdtPosBenchmark  # noqa
 from .fdt_dep import FdtDepBenchmark  # noqa
+from .mim_gold_ner import MimGoldNerBenchmark  # noqa
 from .wikiann_is import WikiannIsBenchmark  # noqa
 from .wikiann_fo import WikiannFoBenchmark  # noqa
 from .norec_is import NorecISBenchmark  # noqa