+++ scandeval/datasets.py
@@ -17,7 +17,7 @@ def load_dataset(name: str) -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -60,7 +60,7 @@ def load_suc3() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -84,7 +84,7 @@ def load_norne_nb() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -108,7 +108,7 @@ def load_norne_nn() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -132,7 +132,7 @@ def load_nordial() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -152,7 +152,7 @@ def load_norec() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -172,7 +172,7 @@ def load_norec_is() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -192,7 +192,7 @@ def load_norec_fo() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -212,7 +212,7 @@ def load_angry_tweets() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -232,7 +232,7 @@ def load_dane() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -256,7 +256,7 @@ def load_ndt_nb_pos() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -280,7 +280,7 @@ def load_ndt_nn_pos() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -304,7 +304,7 @@ def load_ndt_nb_dep() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -328,7 +328,7 @@ def load_ndt_nn_dep() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -352,7 +352,7 @@ def load_ddt_pos() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -376,7 +376,7 @@ def load_ddt_dep() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -400,7 +400,7 @@ def load_dkhate() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -420,7 +420,7 @@ def load_europarl() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -442,7 +442,7 @@ def load_lcc() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -462,7 +462,7 @@ def load_twitter_sent() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -482,7 +482,7 @@ def load_dalaj() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -502,7 +502,7 @@ def load_absabank_imm() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -522,7 +522,7 @@ def load_sdt_pos() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -546,7 +546,7 @@ def load_sdt_dep() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -570,7 +570,7 @@ def load_idt_pos() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -594,7 +594,7 @@ def load_idt_dep() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -612,13 +612,37 @@ def load_idt_dep() -> Tuple[pd.DataFrame, pd.DataFrame,
     return X_train, X_test, y_train, y_test
 
 
+def load_mim_gold_ner() -> Tuple[pd.DataFrame, pd.DataFrame,
+                                 pd.DataFrame, pd.DataFrame]:
+    '''Load the the MIM-GOLD-NER dataset.
+
+    Returns:
+        tuple:
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
+            `X_train` and `X_test` corresponds to the feature matrices for the
+            training and test split, respectively, and `y_train` and `y_test`
+            contains the target vectors.
+    '''
+    base_url = ('https://raw.githubusercontent.com/saattrupdan/ScandEval/'
+                'dev/datasets/mim_gold_ner/')
+    train_url = base_url + 'train.jsonl'
+    test_url = base_url + 'test.jsonl'
+    X_train, y_train = _get_dataset_from_url(train_url,
+                                             ['doc', 'tokens'],
+                                             'ner_tags')
+    X_test, y_test = _get_dataset_from_url(test_url,
+                                           ['doc', 'tokens'],
+                                           'ner_tags')
+    return X_train, X_test, y_train, y_test
+
+
 def load_wikiann_is() -> Tuple[pd.DataFrame, pd.DataFrame,
                                pd.DataFrame, pd.DataFrame]:
     '''Load the the Icelandic WikiANN dataset.
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -642,7 +666,7 @@ def load_wikiann_fo() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -666,7 +690,7 @@ def load_fdt_pos() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.
@@ -690,7 +714,7 @@ def load_fdt_dep() -> Tuple[pd.DataFrame, pd.DataFrame,
 
     Returns:
         tuple:
-            Four dicts, `X_train`, `X_test`, `y_train` and `y_test`, where
+            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
             `X_train` and `X_test` corresponds to the feature matrices for the
             training and test split, respectively, and `y_train` and `y_test`
             contains the target vectors.