+++ CHANGELOG.md
@@ -7,6 +7,11 @@ The format is based on
 and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added the Icelandic NER dataset MIM-GOLD-NER. This can now be loaded as
+  `mim-gold-ner` in the `Benchmark` class and through the CLI.
+
 
 ## [v1.1.3] - 2021-10-04
 ### Fixed